Base project to SwiftUI
