+++ Example/Example/ExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  ExampleApp.swift
+//  Example
+//
+//  Created by João Lucas on 11/11/23.
+//
+
+import SwiftUI
+
+@main
+struct ExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
