+++ Example/Example/Screens/Demo/DemoViewController.swift
@@ -1,0 +1,59 @@
+//
+//  DemoViewController.swift
+//  Example
+//
+//  Created by João Lucas on 11/11/23.
+//
+
+import UIKit
+import SwiftUI
+import MagicImages
+
+struct ViewControllerDemo: UIViewControllerRepresentable {
+        
+    typealias UIViewControllerType = DemoViewController
+    
+    func makeUIViewController(context: Context) -> DemoViewController {
+        DemoViewController()
+    }
+    
+    func updateUIViewController(_ uiViewController: DemoViewController, context: Context) {
+        //
+    }
+    
+}
+
+final class DemoViewController: UIViewController {
+    
+    private let url = "https://images.alphacoders.com/270/270241.jpg"
+    
+    private lazy var firstExample: UIImageView = {
+        let imageView = UIImageView()
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        return imageView
+    }()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        view.backgroundColor = .white
+        
+        view.addSubview(firstExample)
+        
+        firstExample.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
+        firstExample.centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
+        firstExample.widthAnchor.constraint(equalToConstant: 200).isActive = true
+        firstExample.heightAnchor.constraint(equalToConstant: 200).isActive = true
+        
+        MagicImages(image: firstExample).start(url: url, placeholder: "photo")
+        
+        /**
+         MagicImages(image: firstExample).start(name: "photo")
+         
+         OR
+         
+         MagicImages(image: firstExample).start(uiImage: UIImage(systemName: "heart.fill"))
+         */
+    }
+    
+}
